Unify RT shader sample screens into one shared source (#3998)
Collapses the MonoGame/raylib/Skia RenderTargetEffectScreen/RenderTargetShaderScreen
duplicates into one file (RenderTargetShaderScreen.cs) linked into all three sample
csprojs, following the RenderTargetScreen.cs precedent. Platform-necessary bits
(#if RAYLIB/SKIA/else): effect type alias, compile/load body, shader filename, namespace,
color aliasing. Renamed MonoGame's class + nav button from "RT Effect" to "RT Shader" for
consistency (2 of 3 backends already used that name).
