+++ Samples/MonoGameGumInCode/MonoGameGumInCode/MonoGameGumInCode.csproj
@@ -26,7 +26,7 @@
   <ItemGroup>
     <PackageReference Include="MonoGame.Framework.DesktopGL" Version="3.8.4.1" />
     <PackageReference Include="MonoGame.Content.Builder.Task" Version="3.8.4.1" />
-    <!-- Runtime HLSL .fx compiler used by RenderTargetEffectScreen to build a post-process
+    <!-- Runtime HLSL .fx compiler used by RenderTargetShaderScreen to build a post-process
          shader at runtime without the content pipeline (issue #816). -->
     <PackageReference Include="ShadowDusk.Compiler" Version="0.11.0" />
   </ItemGroup>
@@ -49,7 +49,7 @@
     <None Update="Content\**\*.achx">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
-    <!-- Raw .fx text (NOT built through MGCB) — RenderTargetEffectScreen compiles it at runtime
+    <!-- Raw .fx text (NOT built through MGCB) — RenderTargetShaderScreen compiles it at runtime
          via ShadowDusk and references it by path through ContainerRuntime.SourceShaderFile (#3206). -->
     <None Update="Content\**\*.fx">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
